Make a Hashable version of the Collection API
